tests/wpt/web-platform-tests/websockets/Close-Reason-124Bytes.htm

<!DOCTYPE html>
<html>
<head>
    <title>W3C WebSocket API - Close WebSocket - Reason is more than 123 bytes long</title>
    <script type="text/javascript" src="/resources/testharness.js"></script>
    <script type="text/javascript" src="/resources/testharnessreport.js"></script>
    <script type="text/javascript" src="websocket.sub.js"></script>
</head>
<body>
    <div id="log"></div>
    <script type="text/javascript">

        var test = async_test("W3C WebSocket API - Create WebSocket - Close the Connection - close(code, 'reason more than 123 bytes') - SYNTAX_ERR is thrown");

        var wsocket = CreateWebSocket(false, false, false);
        var isOpenCalled = false;

        wsocket.addEventListener('open', test.step_func(function (evt) {
            var reason = "0123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123";
            assert_equals(reason.length, 124);
            assert_throws("SYNTAX_ERR", function () { wsocket.close(1000, reason) });
            test.done();
        }), true);
    </script>

</body>
</html>
